optimizing style

+++ AI-based-Language-Teaching/Service/IQuestionService.cs
@@ -5,9 +5,9 @@ namespace AI_based_Language_Teaching.Service
     public interface IQuestionService
     {
         IEnumerable<Question> GetQuestions();
-        Question GetQuestionById(int id); 
+        Question GetQuestionById(long id); 
         void CreateQuestion(Question question);
         void UpdateQuestion(Question question); 
-        void DeleteQuestion(int id); 
+        void DeleteQuestion(long id); 
     }
 }
